movees NodeConfig and ResearcherConfig modules in component specific folders
adds missing module files
